@@ -1,45 +1,33 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion "28.0.3"
+    namespace 'io.puzzlebox.orbit'
+    compileSdk 36
 
     defaultConfig {
 //        applicationId "info.puzzlebox.orbit"
 //        applicationId "io.puzzlebox.orbit"
-//        minSdkVersion 15 // API 4.0.3
-//        minSdkVersion 16 // API 4.1 [Manifest.permission.READ_EXTERNAL_STORAGE]
-//        minSdkVersion 18 // API 4.3 [Emotiv Community SDK]
-        minSdkVersion 21 // 5.x - required for ScanResult.getDevice()
-        targetSdkVersion 28
+//        minSdk 15 // API 4.0.3
+//        minSdk 16 // API 4.1 [Manifest.permission.READ_EXTERNAL_STORAGE]
+//        minSdk 18 // API 4.3 [Emotiv Community SDK]
+        minSdk 21 // 5.x - required for ScanResult.getDevice()
+        targetSdk 36
         versionCode 22
         versionName "2.4.0"
-        multiDexEnabled true
-        ndk {
-            abiFilters "armeabi"
-        }
+        // No abiFilters: pre-built .so files in jniLibs/armeabi-v7a/ are packaged as-is.
+        // The app runs in 32-bit compat mode on arm64 devices (libNSUART.so and libbedk.so
+        // are armeabi-v7a only; no arm64 builds exist from the SDK vendors).
     }
     buildTypes {
         release {
             minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
-//        debug {
-//            debuggable true
-//        }
-//        staging {
-//            // Specifies a sorted list of fallback build types that the
-//            // plugin should try to use when a dependency does not include a
-//            // "staging" build type. You may specify as many fallbacks as you
-//            // like, and the plugin selects the first build type that's
-//            // available in the dependency.
-//            matchingFallbacks = ['debug', 'qa', 'release']
-//        }
-//        releaseApp {}
-//        releaseSdk {}
     }
-    dexOptions {
-        javaMaxHeapSize "2g"
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
     }
 
     // Disable SerialDevice because not currently used to talk to Puzzlebox Pyramid
