@@ -1,7 +1,7 @@
 package io.puzzlebox.orbit.ui;
 
 import android.os.Bundle;
-import android.support.v7.widget.LinearLayoutCompat;
+import androidx.appcompat.widget.LinearLayoutCompat;
 import android.util.Log;
 import android.view.KeyEvent;
 import android.view.LayoutInflater;
